esg-on-docker/app/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
numpy==1.18.4
psycopg2-binary==2.8.5






